+++ core/server/api_container/main.go
@@ -170,7 +170,6 @@ func createServiceNetworkAndModuleStore(
 	)
 
 	filesArtifactExpander := files_artifact_expander.NewFilesArtifactExpander(
-		args.EnclaveDataDirpathOnHostMachine,
 		kurtosisBackend,
 		enclaveObjAttrsProvider,
 		enclaveId,
@@ -182,7 +181,6 @@ func createServiceNetworkAndModuleStore(
 		kurtosisBackend,
 		filesArtifactExpander,
 		freeIpAddrTracker,
-		args.EnclaveDataDirpathOnHostMachine,
 	)
 
 	networkingSidecarManager := networking_sidecar.NewStandardNetworkingSidecarManager(
@@ -198,13 +196,13 @@ func createServiceNetworkAndModuleStore(
 		kurtosisBackend,
 		enclaveDataDir,
 		userServiceLauncher,
-		networkingSidecarManager)
+		networkingSidecarManager,
+	)
 
 	moduleLauncher := module_launcher.NewModuleLauncher(
 		enclaveId,
 		kurtosisBackend,
 		apiContainerSocketInsideNetwork,
-		args.EnclaveDataDirpathOnHostMachine,
 		freeIpAddrTracker,
 	)